+++ src/components/Image/Image.tsx
@@ -4,6 +4,9 @@ import { GradeBand } from 'src/enum/gradeband';
 export type ImageProps = {
   imageSrc: string;
   altText: string;
+  roundedCorners?: boolean;
+  contain?: boolean;
+  className?: string;
   dataTestId?: string;
   gradeBand?: GradeBand;
 };
@@ -15,6 +18,24 @@ const defaultImageSrc =
     ? '/node_modules/@connect/connect-design-system/dist/images/zelda.jpg'
     : '/images/zelda.jpg';
 
-export const Image: React.FC<ImageProps> = ({ imageSrc, altText, dataTestId }) => {
-  return <img src={imageSrc || defaultImageSrc} alt={altText} data-testid={dataTestId} />;
+export const Image: React.FC<ImageProps> = ({
+  imageSrc,
+  altText,
+  roundedCorners,
+  contain,
+  className,
+  dataTestId,
+}) => {
+  const classNames = ['connect__image', roundedCorners && 'connect__rounded-corners', contain && 'connect__contain', className]
+    .filter(Boolean)
+    .join(' ');
+
+  return (
+    <img
+      className={classNames}
+      src={imageSrc || defaultImageSrc}
+      alt={altText}
+      data-testid={dataTestId}
+    />
+  );
 };